+++ packages/contracts/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- add natspec docs to balance manager
+- remove admin prefill logic and replace with new scheme were handling refunds attempts to reserve 1 erc20 token each time if balance == 0
 - fix invariant test address setup so all addrs come from `InvariantsBase.sol`
 - add natspec docs to CTM
 - move fill batch logic and state from handler into CTM (we already expose applySubtreeUpdate externally, all the related fillbatch logic should remain in the same contract)